# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: ByteDanceDemo API
  description: |-
    A simplified TikTok/Douyin clone backend API providing video sharing,
    user interactions, and social features.
  version: 1.0.0
  contact:
    name: ByteDanceDemo Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
  /douyin/feed/:
    get:
      tags:
        - Video
      summary: Get video feed
      description: Returns list of videos with optional pagination and user authentication
      operationId: getFeed
      parameters:
        - name: token
          in: query
          description: User authentication token (optional)
          required: false
          schema:
            type: string
        - name: latest_time
          in: query
          description: Optional timestamp for pagination
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful feed retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
  /douyin/user/:
    get:
      tags:
        - User
      summary: Get user information
      description: Returns detailed user information
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          description: User ID to retrieve (defaults to current user)
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful user info retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /douyin/user/register/:
    post:
      tags:
        - User
      summary: User registration
      description: Register a new user account
      parameters:
        - name: username
          in: query
          description: Username (3-20 characters, alphanumeric)
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 20
        - name: password
          in: query
          description: Password (min 6 characters, must contain letters and numbers)
          required: true
          schema:
            type: string
            minLength: 6
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /douyin/user/login/:
    post:
      tags:
        - User
      summary: User login
      description: Authenticate user and return access token
      parameters:
        - name: username
          in: query
          description: Username
          required: true
          schema:
            type: string
        - name: password
          in: query
          description: Password
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserLoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BasicResponse:
      type: object
      properties:
        status_code:
          type: integer
          format: int32
          description: Status code (0=success, 1=error)
        status_msg:
          type: string
          description: Status message
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: User ID
        name:
          type: string
          description: Username
        follow_count:
          type: integer
          format: int64
          description: Number of following
        follower_count:
          type: integer
          format: int64
          description: Number of followers
        is_follow:
          type: boolean
          description: Whether current user follows this user
    Video:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Video ID
        author:
          $ref: '#/components/schemas/User'
          description: Video author information
        play_url:
          type: string
          description: Video playback URL
        cover_url:
          type: string
          description: Video cover image URL
        favorite_count:
          type: integer
          format: int64
          description: Number of likes
        comment_count:
          type: integer
          format: int64
          description: Number of comments
        is_favorite:
          type: boolean
          description: Whether current user liked this video
...
    Response:
      type: object
      properties:
        status_code:
          type: integer
          format: int32
          description: Status code (0=success, 1=error)
        status_msg:
          type: string
          description: Status message
    FeedResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            video_list:
              type: array
              items:
                $ref: '#/components/schemas/Video'
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    UserLoginResponse:
      allOf:
        - $ref: '#/components/schemas/Response'
        - type: object
          properties:
            user_id:
              type: integer
              format: int64
              description: User ID
            token:
              type: string
              description: JWT authentication token
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BasicResponse'